backend/pyproject.toml
[project]
name = "cooking-backend"
version = "0.1.0"
description = "Backend orchestrator for the voice-first cooking & health iOS app"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "structlog>=24.4.0",
    "sse-starlette>=2.1.3",
    "anyio>=4.6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "types-python-jose>=3.3.4",
    "types-passlib>=1.7.7",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "ASYNC"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["passlib.*", "jose.*", "sse_starlette.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
addopts = "-ra"
